requirements.txt
Flask==2.0.1
SQLAlchemy==1.4.23
boto3==1.18.44
Werkzeug==2.0.1
python-magic==0.4.24
python-multipart==0.0.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
sqlalchemy>=2.0.0
aiofiles>=0.7.0
webdavclient3>=3.14.6
python-dotenv>=1.0.0
pydantic>=2.4.2
openai>=1.3.0
numpy>=1.26.0
alembic>=1.12.1
pydantic-settings>=2.2.1
fastapi==0.68.1
uvicorn==0.15.0 